Adjust mutations to accept array of variableDeclarations
I had a heck of a time getting the basic exmaple in the spec working in this lib.  I think a sensible approach
is to allow for an array of arguments (Variable Declarations) instead of a wrapper object with a 'key'.  I've implemented
a test case that follows the example in the spec.

https://facebook.github.io/graphql/#sec-Language.Operations